aergo-lambda-cli/requirements.txt
astroid==2.2.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colored==1.3.93
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauth2client==4.1.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
PyDrive==1.3.1
pylint==2.3.1
pyoauth2==0.0.4
python-gitlab==1.11.0
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.5
wrapt==1.11.2
